move loading code to loader
